handle multiple plays run
In one "run" we can pass many playbooks, for example:

```
$ ansible-playbook playbook1.yml playbook2.yml
```

Or many "plays" in on playbook file:

```yaml
$ cat playbook.yml
---
- name: "PLAY 1"
  hosts: "all"
  tasks:
    - name: "ping 1"
      ping:

- name: "PLAY 2"
  hosts: "all"
  tasks:
    - name: "ping 2"
      ping:
```

This callback plugin open only one trace file for an entire "run", but
this file is closed at the end of the first playbook or play (in
`v2_playbook_on_stats`), and thus it raise an "already close" exception
for the second playbook/play.

This change close the file at the end of the "run".